backend/requirements.txt
fastapi==0.115.9
fastapi-cli==0.0.7
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-experimental==0.3.4
langchain-neo4j==0.4.0
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
neo4j==5.28.1
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
uvicorn==0.34.2